fix: setSendMax of null error
```
(node:26) UnhandledPromiseRejectionWarning: TypeError: Cannot read property 'setSendMax' of null
    at Payer.pay (/app/node_modules/web-monetization-receiver/src/payer.js:42:12)
    at process._tickCallback (internal/process/next_tick.js:68:7)
```

If `pay()` is called twice, but the second call occurs before the STREAM connection is established, `receiver.getStream()` would return `null`. This patch stores a promise to the receiver in `receivers` instead of the receiver. The promise resolves when the connection is established.